backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4

# AWS SDK and services
boto3==1.34.144
botocore==1.34.144

# Authentication and security
PyJWT==2.8.0
cryptography==42.0.8
python-multipart==0.0.9

# HTTP client for Cognito JWKS
requests==2.32.3

# Audio processing libraries
pydub==0.25.1
mutagen==1.47.0
librosa==0.10.2
scipy==1.13.1
numpy==1.26.4
soundfile==0.12.1

# AWS Lambda integration
mangum==0.17.0

# JWT handling for Cognito
python-jose[cryptography]==3.3.0

# Database ORM
sqlalchemy==2.0.41
alembic==1.13.1

# Data validation and utilities
python-dateutil==2.9.0.post0
email-validator==2.2.0

# Development and testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0

# Linting and formatting
black==24.4.2
isort==5.13.2
flake8==7.1.0
mypy==1.10.1

# Documentation
mkdocs==1.6.0
mkdocs-material==9.5.27

# Environment management
python-dotenv==1.0.1

# Production server
gunicorn==22.0.0

# Monitoring and observability
structlog==24.2.0

# Additional production dependencies
slowapi==0.1.9  # Better rate limiting
redis==5.0.7  # For distributed caching and rate limiting